Moves syncBotConfigs function to new gambit-junior lib
