+++ api/router.js
@@ -7,6 +7,7 @@ var reportbackRouter = require('./legacy/reportback');
 var CampaignBot = require('./controllers/CampaignBotController');
 var DonorsChooseBot = require('./controllers/DonorsChooseBotController');
 var Slothbot = require('./controllers/SlothBotController');
+var gambitJunior = rootRequire('lib/gambit-junior');
 
 app.use('/', router);
 
@@ -55,7 +56,6 @@ router.post('/v1/chatbot', function(request, response) {
 
 router.post('/v1/chatbot/sync', function(request, response) {
 
-  var controller = new DonorsChooseBot();
-  controller.syncBotConfigs(request, response);
+  gambitJunior.syncBotConfigs(request, response, request.query.bot_type);
 
 });
